# Fix bugs in PhpArrayAdapter introduction