--- a/DependencyInjection/Configuration.php
+++ b/DependencyInjection/Configuration.php
@@ -688,6 +688,13 @@ class Configuration implements ConfigurationInterface
                 ->scalarNode('pool')->end()
             ->end();
 
+        if ($name === 'metadata_cache_driver') {
+            $node->setDeprecated(...$this->getDeprecationMsg(
+                'The "metadata_cache_driver" configuration key is deprecated. PHP Array cache is now automatically registered when %kernel.debug% is false.',
+                '2.3'
+            ));
+        }
+
         return $node;
     }