--- a/Tests/DependencyInjection/DoctrineExtensionTest.php
+++ b/Tests/DependencyInjection/DoctrineExtensionTest.php
@@ -2,6 +2,7 @@
 
 namespace Doctrine\Bundle\DoctrineBundle\Tests\DependencyInjection;
 
+use Doctrine\Bundle\DoctrineBundle\CacheWarmer\DoctrineMetadataCacheWarmer;
 use Doctrine\Bundle\DoctrineBundle\DependencyInjection\DoctrineExtension;
 use Doctrine\Bundle\DoctrineBundle\Tests\Builder\BundleConfigurationBuilder;
 use Doctrine\DBAL\Connection;
@@ -12,6 +13,8 @@ use LogicException;
 use PHPUnit\Framework\TestCase;
 use Symfony\Bridge\Doctrine\Messenger\DoctrineClearEntityManagerWorkerSubscriber;
 use Symfony\Component\Cache\Adapter\ArrayAdapter;
+use Symfony\Component\Cache\Adapter\DoctrineAdapter;
+use Symfony\Component\Cache\Adapter\PhpArrayAdapter;
 use Symfony\Component\Cache\DoctrineProvider;
 use Symfony\Component\DependencyInjection\Compiler\ResolveChildDefinitionsPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -416,7 +419,29 @@ class DoctrineExtensionTest extends TestCase
         $this->assertEquals('doctrine.orm.quote_strategy.default', (string) $calls[11][1][0]);
         $this->assertEquals('doctrine.orm.default_entity_listener_resolver', (string) $calls[12][1][0]);
 
+        $definition = $container->getDefinition('doctrine.orm.default_metadata_cache_warmer');
+        $this->assertSame(DoctrineMetadataCacheWarmer::class, $definition->getClass());
+        $this->assertEquals(
+            [
+                new Reference('doctrine.orm.default_entity_manager'),
+                '%kernel.cache_dir%/doctrine/orm/default_metadata.php',
+            ],
+            $definition->getArguments()
+        );
+
         $definition = $container->getDefinition((string) $container->getAlias('doctrine.orm.default_metadata_cache'));
+        $this->assertEquals(DoctrineProvider::class, $definition->getClass());
+        $arguments = $definition->getArguments();
+        $this->assertInstanceOf(Definition::class, $arguments[0]);
+        $this->assertEquals(PhpArrayAdapter::class, $arguments[0]->getClass());
+        $arguments = $arguments[0]->getArguments();
+        $this->assertSame('%kernel.cache_dir%/doctrine/orm/default_metadata.php', $arguments[0]);
+        $this->assertInstanceOf(Definition::class, $arguments[1]);
+        $this->assertEquals(DoctrineAdapter::class, $arguments[1]->getClass());
+        $arguments = $arguments[1]->getArguments();
+        $this->assertInstanceOf(Reference::class, $arguments[0]);
+        $this->assertEquals('doctrine.orm.cache.provider.cache.doctrine.orm.default.metadata', (string) $arguments[0]);
+        $definition = $container->getDefinition((string) $arguments[0]);
         $this->assertEquals(DoctrineProvider::class, $definition->getClass());
         $arguments = $definition->getArguments();
         $this->assertInstanceOf(Reference::class, $arguments[0]);
@@ -824,6 +849,9 @@ class DoctrineExtensionTest extends TestCase
         }
     }
 
+    /**
+     * @group legacy
+     */
     public function testInvalidCacheConfiguration(): void
     {
         if (! interface_exists(EntityManagerInterface::class)) {
@@ -870,15 +898,42 @@ class DoctrineExtensionTest extends TestCase
         $this->assertEquals($expectedAliasTarget, (string) $alias);
     }
 
-    public static function cacheConfigurationProvider(): array
+    /**
+     * @dataProvider legacyCacheConfigurationProvider
+     * @group legacy
+     */
+    public function testLegacyCacheConfiguration(string $expectedAliasName, string $expectedAliasTarget, string $cacheName, array $cacheConfig): void
+    {
+        $this->testCacheConfiguration($expectedAliasName, $expectedAliasTarget, $cacheName, $cacheConfig);
+    }
+
+    public static function legacyCacheConfigurationProvider(): array
     {
         return [
             'metadata_cache_default' => [
                 'expectedAliasName' => 'doctrine.orm.default_metadata_cache',
-                'expectedAliasTarget' => 'doctrine.orm.cache.provider.cache.doctrine.orm.default.metadata',
+                'expectedAliasTarget' => 'doctrine.orm.cache.provider.cache.doctrine.orm.default.metadata.php_array',
                 'cacheName' => 'metadata_cache_driver',
                 'cacheConfig' => ['type' => null],
             ],
+            'metadata_cache_pool' => [
+                'expectedAliasName' => 'doctrine.orm.default_metadata_cache',
+                'expectedAliasTarget' => 'doctrine.orm.cache.provider.metadata_cache_pool.php_array',
+                'cacheName' => 'metadata_cache_driver',
+                'cacheConfig' => ['type' => 'pool', 'pool' => 'metadata_cache_pool'],
+            ],
+            'metadata_cache_service' => [
+                'expectedAliasName' => 'doctrine.orm.default_metadata_cache',
+                'expectedAliasTarget' => 'service_target_metadata.php_array',
+                'cacheName' => 'metadata_cache_driver',
+                'cacheConfig' => ['type' => 'service', 'id' => 'service_target_metadata'],
+            ],
+        ];
+    }
+
+    public static function cacheConfigurationProvider(): array
+    {
+        return [
             'query_cache_default' => [
                 'expectedAliasName' => 'doctrine.orm.default_query_cache',
                 'expectedAliasTarget' => 'doctrine.orm.cache.provider.cache.doctrine.orm.default.query',
@@ -891,13 +946,6 @@ class DoctrineExtensionTest extends TestCase
                 'cacheName' => 'result_cache_driver',
                 'cacheConfig' => ['type' => null],
             ],
-
-            'metadata_cache_pool' => [
-                'expectedAliasName' => 'doctrine.orm.default_metadata_cache',
-                'expectedAliasTarget' => 'doctrine.orm.cache.provider.metadata_cache_pool',
-                'cacheName' => 'metadata_cache_driver',
-                'cacheConfig' => ['type' => 'pool', 'pool' => 'metadata_cache_pool'],
-            ],
             'query_cache_pool' => [
                 'expectedAliasName' => 'doctrine.orm.default_query_cache',
                 'expectedAliasTarget' => 'doctrine.orm.cache.provider.query_cache_pool',
@@ -909,13 +957,6 @@ class DoctrineExtensionTest extends TestCase
                 'expectedAliasTarget' => 'doctrine.orm.cache.provider.result_cache_pool',
                 'cacheName' => 'result_cache_driver',
                 'cacheConfig' => ['type' => 'pool', 'pool' => 'result_cache_pool'],
-            ],
-
-            'metadata_cache_service' => [
-                'expectedAliasName' => 'doctrine.orm.default_metadata_cache',
-                'expectedAliasTarget' => 'service_target_metadata',
-                'cacheName' => 'metadata_cache_driver',
-                'cacheConfig' => ['type' => 'service', 'id' => 'service_target_metadata'],
             ],
             'query_cache_service' => [
                 'expectedAliasName' => 'doctrine.orm.default_query_cache',