Second attempt at doing #1196

Fixes:
- #1251 by deprecating clear metadata cache command
- #1249 by adding priority for cache warmer
- #1254 thank you @dmaicher