@ossinkine 2020-11-13
When I deprecate the command by adding `<deprecated>` tag to XML-config, the tests are failing with an error `Since symfony/dependency-injection 5.1: Not setting the attribute "package" of the node "deprecated" in "/home/travis/build/doctrine/DoctrineBundle/DependencyInjection/../Resources/config/orm.xml" is deprecated.
` (https://travis-ci.com/github/doctrine/DoctrineBundle/builds/201245230)
When I add this attributes, other tests are failing with an error `Symfony\Component\DependencyInjection\Exception\InvalidArgumentException: Unable to parse file "/home/travis/build/doctrine/DoctrineBundle/DependencyInjection/../Resources/config/orm.xml": [ERROR 1827] Element '{http://symfony.com/schema/dic/services}deprecated', attribute 'package': The attribute 'package' is not allowed. (in /home/travis/build/doctrine/DoctrineBundle/ - line 156, column 0)
` (https://travis-ci.com/github/doctrine/DoctrineBundle/builds/201255247)