@ostrolucky 2020-11-13
Yeah, this is why we don't usually deprecate commands in xml, just put `@deprecated` and trigger_error inside command, similarly as is done in other commands