@greg0ire 2020-11-15
Please do one commit per bug fixed, and for each bugfix, summarize what's broken, and explain why your changes fix what is broken.