@ossinkine 2020-11-15
@greg0ire I've updated the commits. I've put the links to fixed issues in started post, they explaining what's broken and how it proposed to fix.