@greg0ire 2020-11-15
>  I've put the links to fixed issues in started post, they explaining what's broken and how it proposed to fix.

That's my point, you shouldn't have to read lengthy discussions. Since you insist on not doing this, then I'm going to do it myself for the next reviewer, or for the person who is going to read the history if these commits somehow cause an issue. This is going to take me some time, but I can't give a good review if I fully understand what this is about first.