@ossinkine 2020-11-15
> That's my point, you shouldn't have to read lengthy discussions. Since you insist on not doing this, then I'm going to do it myself for the next reviewer, or for the person who is going to read the history if these commits somehow cause an issue. This is going to take me some time, but I can't give a good review if I fully understand what this is about first.

IMO if you want to figure out why changes was made you whatever will read all related to understand the root of the issue, what initial state produce the issue and so on, even if the problem is described here.