@ossinkine 2020-11-15
> The first commit is labeled "Add PhpArrayAdapter to cache metadata", but that's the title of #1196 … you are not doing something twice, are you?

#1196 was reverted in #1255
I just commited it again