@greg0ire 2020-11-15
Shouldn't the first commit be squashed with the first one then? What the point of introducing something broken then fixing it in the commit right after?