@ossinkine 2020-11-15
> Please do one commit per bug fixed

🤷‍♂️