@greg0ire 2020-11-15
All the bugs this PR are pointing to are fixed by the revert, aren't they? Or can some of the bugs still be experienced with the latest version?