@greg0ire 2020-11-15
About the cache metadata clearing command, it's deprecated, but running it will still cause an issue, won't it? In short, if we merge and release this, can everybody just run `composer update` and their deploy scripts will still work, or not?