@ossinkine 2020-11-15
> All the bugs this PR are pointing to are fixed by the revert, aren't they? Or can some of the bugs still be experienced with the latest version?

Yes, all bugs was fixed by the revert, because they was introduced by inital PR.
This PR contains the initial commit and the bug fixes which the inital commit has.