@ossinkine 2020-11-15
> About the cache metadata clearing command, it's deprecated, but running it will still cause an issue, won't it? In short, if we merge and release this, can everybody just run `composer update` and their deploy scripts will still work, or not?

`composer update` does not execute the metadata cache clearing by default, so it's safe. If somebody has the metadata cache clearing in his post-install scripts, he should remove it because it does not make sence now.