@greg0ire 2020-11-15
> UPD: This only applies to the no debug mode (usually only prod environment)

That makes the BC-break far worse because it cannot be detected in dev.