@greg0ire 2020-11-15
> Yes, all bugs was fixed by the revert, because they was introduced by inital PR. This PR contains the initial commit and the bug fixes which the inital commit has.

Then it's simple: do one commit. If you have a lot of explaining to do, do it in the body of the commit message, or with code comments.