@ossinkine 2020-11-15
> Ok so it's a BC-break I'm afraid. You can be sure you will get bug reports from people about this. But since this is only an issue with the new cache system you are introducing, maybe you can somehow make the command always succeed in addition to the deprecation?

I can add a check to the command if PHP Array cache is enabled we just return a success result with the appropriate message