@dmaicher 2020-11-15
I think we should do some more benchmarks and make sure this change is actually worth it?

I just tried it on one of my apps and I can actually not measure any real performance improvement.

I changed my front controller to just this:

```php
require __DIR__.'/../vendor/autoload.php';

$kernel = new AppKernel('prod', false);
$kernel->boot();

/** @var \Doctrine\ORM\EntityManagerInterface $em */
$em = $kernel->getContainer()->get('doctrine.orm.default_entity_manager');
$em->getMetadataFactory()->getAllMetadata();
``` 

I made sure to clear the cache and re-start php-fpm in between the runs.

I got those results on 3 runs with 10000 requests each using apache-bench

```
ab -n 100000 -c 1 "http://my.app"
``` 

And I'm taking `Time per request (mean)` as the benchmark here.

| cache type | run 1 | run 2 | run 3 | 
|---|---|---|---|
| old meta data caching using `cache.system` | 0.203ms  | 0.206ms | 0.204ms |  
| this PR | 0.202ms  | 0.205ms  | 0.203ms  |  


Some details on my app and system:
- Symfony 4.4
- PHP 7.4.12
- 116 Entity classes on default EM. So `getAllMetadata()` loaded data for all 116 (checked that).

@ossinkine can you share some details on the benchmarks you did?