@dmaicher 2020-11-15
I just realized my first benchmark numbers shared above were wrong. I had Varnish in front of nginx and php-fpm and it was just returning cached responses. :man_facepalming:

Now the numbers are updated and correct.