@ossinkine 2020-11-16
Unfortunatelly I did not do any synthetic tests but I just apply this PR on my prod app and I get the following results in the monitoring:
![Screenshot_1](https://user-images.githubusercontent.com/2588533/99246161-a3fbf980-2826-11eb-937c-837e1b1b7f66.png)

Could you please try to benchmark concurrent requests, for example using `-c 10` flag?