@ossinkine 2020-11-16
Median response time with APCu cache is about 84ms but with PHP array cache is about 80ms.
The system configuration like you: PHP 7.4.12, Symfony 4.4, about 80 entities.