@dmaicher 2020-11-16
@ossinkine interesting. And what cache configuration did you use exactly before the deploy of the change? 

I used this "before" config: https://github.com/symfony/recipes/blob/master/doctrine/doctrine-bundle/1.12/config/packages/prod/doctrine.yaml#L6

Let me try to increase the concurrency and do another benchmark but I don't see how this would affect the results