@ostrolucky 2020-11-16
you need to be sure you have opcache correctly configured as well