@ossinkine 2020-12-28
Ok, I made exactly the same benchmark on my localhost app as @dmaicher described and there is my results below:

cache type | run 1 | run 2 | run 3
-- | -- | -- | --
old meta data caching using cache.system | 16.619ms | 16.891ms | 15.958ms
this PR | 14.576ms | 14.535ms | 14.747ms

- Symfony 4.4.18
- PHP 7.4.13
- 87 entities